Revert "Gopi Update" 13.20
This reverts commit 5d389858b427cdf44d9b6de5ba73e7057bf7f96a.

+++ HomeAppLoanApplication/src/main/java/com/homeloan/main/retalitionalmodel/EMIDetails.java
@@ -16,7 +16,7 @@ public class EMIDetails {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long emiID;
+	private Integer emiID;
 	private String nextEmiDueDate,
 	previousEmiStatus;
 	private Double emiAmountMonthly;
